Refactoring + add documentation
