+++ src/components/MxGraph.tsx
@@ -4,7 +4,7 @@ import {ModelComponent, ModelTaskMetadata} from "../entities/modelTaskMetadata";
 import {DataAccessProvider} from "../persistence/dataAccessProvider";
 import {GetValidationReport, VisualizeComponent} from "../services/ModelTasks";
 import TimingService from "../services/TimingService";
-import {ValidationReport} from "../conformance/wrapper/ValidationReport";
+import {ValidationReport} from "../conformance/ValidationReport";
 import {MxGraphProps} from "./interfaces/interfaces";
 
 import {ModelObserver} from "../entities/model";
