+++ src/persistence/dataAccessProvider.ts
@@ -2,6 +2,9 @@ import {Model} from "../entities/model";
 import {FileDAO} from "./fileDAO";
 import {ValidationService} from "../conformance/ValidationService";
 
+/**
+ * This class provides access to most of the application's key components.
+ */
 export class DataAccessProvider {
 
     private static _instance: DataAccessProvider = new DataAccessProvider();
@@ -11,6 +14,9 @@ export class DataAccessProvider {
     // tmp field
     private _model: Model;
 
+    /**
+     * Create a new Data Access Provider.
+     */
     private constructor() {
         // temporarily create model here
         this._model = new Model();
@@ -19,15 +25,26 @@ export class DataAccessProvider {
         this.validationService = new ValidationService(this._model);
     }
 
-    // tmp method
+    /**
+     * Retrieve the model.
+     * @returns {Model}
+     */
     get model(): Model {
         return this._model;
     }
 
+    /**
+     * Retrieve an instance of this data access provider.
+     * @returns {DataAccessProvider}
+     */
     public static getInstance(): DataAccessProvider {
         return this._instance;
     }
 
+    /**
+     * Retrieve the file DAO.
+     * @returns {FileDAO}
+     */
     public getFileDAO(): FileDAO {
         if (this.fileDAO) {
             return this.fileDAO;
@@ -37,6 +54,10 @@ export class DataAccessProvider {
         }
     }
 
+    /**
+     * Retrieve the validation service.
+     * @returns {ValidationService}
+     */
     public getValidationService(): ValidationService {
         return this.validationService;
     }
